Change how pushes work internally, pop will use the correct type of push, add a generic sliding section view controller

+++ src/ViewImprovements/AB_TappableTextView.m
@@ -105,7 +105,7 @@
 
 - (void) showLabel
 {
-    characterCountLabel.text = [NSString stringWithFormat:@"%d/%d", self.text.length, _maxCharacterCount];
+    characterCountLabel.text = [NSString stringWithFormat:@"%lu/%d", (unsigned long)self.text.length, _maxCharacterCount];
 }
 
 - (void)textViewDidChange:(UITextView *)textView
